refactor: move terminate tests too separate class
